refactor: move theme utils and config to correct locations
